added extension lowering and [dry run] mode

+++ src/main.cpp
@@ -12,10 +12,25 @@ std::unordered_map<std::string, std::string> fileCategories {
     {".exe", "Executable"}
 };
 
+std::string lower(std::string s) {
+    std::string newS;
+    for (char ch : s) {
+        newS.push_back(std::tolower(ch));
+    }
+    return newS;
+}
+
 int main(int argc, char* argv[]) {
     if (argc < 2) {
         std::cerr << "Usage: organizer.exe <directory>\n";
         return 1;
+    }
+
+    bool dryRun = false;
+    for (int i = 2; i < argc; i++) {
+        if (std::string(argv[i]) == "--dry-run") {
+            dryRun = true;
+        }
     }
 
     fs::path directory = argv[1];
@@ -34,28 +49,33 @@ int main(int argc, char* argv[]) {
     int count{};
     int errors{};
     int unknown{};
+
+    if (dryRun) std::cout << "[DRY RUN]\n";
     
     for (const auto& entry : fs::directory_iterator(directory)) {
         if (!entry.is_regular_file()) {
             continue;
         }
 
-        auto it = fileCategories.find(entry.path().extension().string());
+        std::string extension = entry.path().extension().string();
+        auto it = fileCategories.find(lower(extension));
         if (it != fileCategories.end()) {
-            try {
-                auto destinationPath = directory / it->second;
-                if (!fs::exists(destinationPath)) {
-                    fs::create_directory(destinationPath);
-                }
+            if (!dryRun) {
+                try {
+                    auto destinationPath = directory / it->second;
+                    if (!fs::exists(destinationPath)) {
+                        fs::create_directory(destinationPath);
+                    }
 
-                fs::rename(entry.path(), destinationPath / entry.path().filename());
-            }
-            catch (const fs::filesystem_error& e) {
-                std::cerr << "Filesystem error: " << e.what() << '\n';
-                std::cerr << "Source path: " << e.path1() << '\n';
-                std::cerr << "Destination path: " << e.path2() << '\n';
-                errors++;
-                continue;
+                    fs::rename(entry.path(), destinationPath / entry.path().filename());
+                }
+                catch (const fs::filesystem_error& e) {
+                    std::cerr << "Filesystem error: " << e.what() << '\n';
+                    std::cerr << "Source path: " << e.path1() << '\n';
+                    std::cerr << "Destination path: " << e.path2() << '\n';
+                    errors++;
+                    continue;
+                }
             }
 
             std::cout << "Moved: " << entry.path().filename() << " -> " << it->second << '\n';
@@ -67,9 +87,11 @@ int main(int argc, char* argv[]) {
         }
     }
 
-    std::cout << "Moved files: " << count << '\n'
-        << "Unknown files: " << unknown << '\n'
-        << "Errors: " << errors << '\n'; 
+    if (!dryRun) {
+        std::cout << "Moved files: " << count << '\n'
+            << "Unknown files: " << unknown << '\n'
+            << "Errors: " << errors << '\n'; 
+    }
 
     return 0;
 }
